reuse pubKeyFromMnemonics
