+++ lib/core-integration/src/Test/Integration/Scenario/API/HWWallets.hs
@@ -28,13 +28,7 @@ import Cardano.Wallet.Api.Types
     , WalletStyle (..)
     )
 import Cardano.Wallet.Primitive.AddressDerivation
-    ( FromMnemonic (..)
-    , HardDerivation (..)
-    , PersistPublicKey (..)
-    , WalletKey (..)
-    )
-import Cardano.Wallet.Primitive.AddressDerivation.Shelley
-    ( generateKeyFromSeed )
+    ( NetworkDiscriminant (..) )
 import Cardano.Wallet.Primitive.AddressDiscovery.Sequential
     ( defaultAddressPoolGap, getAddressPoolGap )
 import Cardano.Wallet.Primitive.Mnemonic
@@ -75,6 +69,7 @@ import Test.Integration.Framework.DSL
     , listAddresses
     , mkEpochInfo
     , notDelegating
+    , pubKeyFromMnemonics
     , quitStakePool
     , request
     , selectCoins
@@ -88,7 +83,6 @@ import Test.Integration.Framework.TestData
     ( errMsg403NoRootKey, payloadWith, updateNamePayload, updatePassPayload )
 
 import qualified Cardano.Wallet.Api.Link as Link
-import qualified Data.Text.Encoding as T
 import qualified Network.HTTP.Types.Status as HTTP
 
 
@@ -435,14 +429,6 @@ spec = do
                 ]
 
  where
-     pubKeyFromMnemonics :: [Text] -> Text
-     pubKeyFromMnemonics mnemonics =
-         T.decodeUtf8 $ serializeXPub $ publicKey
-            $ deriveAccountPrivateKey mempty rootXPrv minBound
-      where
-          (Right seed) = fromMnemonic @'[15] mnemonics
-          rootXPrv = generateKeyFromSeed (seed, Nothing) mempty
-
      restoredWalletName :: Text
      restoredWalletName = "Wallet from pub key"
